scripts/packagesynccommon: Deploy s3indexbuilder.py. Also make sure we currently sync only x86_64 packages to yonada. The rest is pulled from yonada:

+++ scripts/buildserver/packagesynccommon.sh
@@ -42,17 +42,22 @@ echo $GPG_PASSWORD | /usr/bin/gpg2 -a --pinentry-mode loopback --detach-sign --b
 echo $GPG_PASSWORD | /usr/bin/gpg2 -a --pinentry-mode loopback --detach-sign --batch --yes --passphrase-fd 0 $COMMON_SRPM_DIR/repodata/repomd.xml
 echo $GPG_PASSWORD | /usr/bin/gpg2 -a --pinentry-mode loopback --detach-sign --batch --yes --passphrase-fd 0 $COMMON_DEBUG_RPM_DIR/repodata/repomd.xml
 
-# Finally, perform the rsync:
-
-rsync --checksum -ave ssh --delete $COMMON_RPM_DIR/ yumupload@yum.postgresql.org:yum/yum/common/$osdistro/$os-$osarch
-rsync --checksum -ave ssh --delete $COMMON_SRPM_DIR/ yumupload@yum.postgresql.org:yum/yum/srpms/common/$osdistro/$os-$osarch
+# We currently sync only x86_64 packages to yonada. The rest is pulled from yonada:
+if [ "$osarch" = "x86_64" ]
+then
+	# Finally, perform the rsync:
+	rsync --checksum -ave ssh --delete $COMMON_RPM_DIR/ yumupload@yum.postgresql.org:yum/yum/common/$osdistro/$os-$osarch
+	rsync --checksum -ave ssh --delete $COMMON_SRPM_DIR/ yumupload@yum.postgresql.org:yum/yum/srpms/common/$osdistro/$os-$osarch
+fi
 
 # Sync SRPMs to S3 bucket:
 aws s3 sync $COMMON_SRPM_DIR s3://dnf-srpms.postgresql.org20250313103537584600000001/srpms/common/$osdistro/$os-$osarch/ --exclude "*.html"
+~/bin/s3indexbuilder.py dnf-srpms.postgresql.org20250313103537584600000001 srpms/common/$osdistro/$os-$osarch --cfdistribution $CF_SRPM_DISTRO_ID
 aws cloudfront create-invalidation --distribution-id $CF_SRPM_DISTRO_ID --path /srpms/common/$osdistro/$os-$osarch/repodata/*
 
 # Sync debug* RPMs to S3 bucket:
 aws s3 sync $COMMON_DEBUG_RPM_DIR s3://dnf-debuginfo.postgresql.org20250312201116649700000001/debug/common/$osdistro/$os-$osarch/ --exclude "*.html"
+~/bin/s3indexbuilder.py dnf-debuginfo.postgresql.org20250312201116649700000001 debug/common/$osdistro/$os-$osarch --cfdistribution $CF_DEBUG_DISTRO_ID
 aws cloudfront create-invalidation --distribution-id $CF_DEBUG_DISTRO_ID --path /debug/common/$osdistro/$os-$osarch/repodata/*
 
 exit 0
